pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "french-assistant"
version = "0.1.0"
description = "RAG-система для помощи в переводе с русского на французский язык"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "French Assistant Team"}
]
keywords = [
    "rag",
    "langchain",
    "nlp",
    "translation",
    "french",
    "russian",
    "llm",
    "chromadb"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-text-splitters>=0.2.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
]
huggingface = [
    "langchain-huggingface>=0.0.3",
    "accelerate>=0.25.0",
    "bitsandbytes>=0.41.0",
]
all = [
    "french-assistant[dev,huggingface]",
]

[project.urls]
Homepage = "https://github.com/MaxainNN/french-assistant"
Documentation = "https://github.com/MaxainNN/french-assistant#readme"
Repository = "https://github.com/MaxainNN/french-assistant"
Issues = "https://github.com/MaxainNN/french-assistant/issues"

[project.scripts]
french-assistant = "french_assistant.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
french_assistant = ["py.typed", "default_config.yaml"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
    | \.venv
    | venv
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip = [".git", "venv", ".venv"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
